+++ test/e2e/specs/autoNumeric.e2e.spec.js
@@ -61,23 +61,26 @@ const selectors = {
     issue303inputNonAn    : '#issue_303non_an',
     issue303inputP        : '#issue_303p',
     issue303inputS        : '#issue_303s',
+    issue403a             : '#issue_403a',
+    issue403b             : '#issue_403b',
+    issue403c             : '#issue_403c',
 };
 
 //-----------------------------------------------------------------------------
 // ---- Helper functions
 
 /*
-function helperGetCaretPosition(wdioElement) { //FIXME Find a way to allow using helper functions inside webdriver.io `execute()` blocks
-    console.log('wdioElement:', wdioElement); //DEBUG 
-    // console.log('this:', this); //DEBUG
-    const selector = wdioElement.selector;
-    console.log('selector:', selector); //DEBUG 
+ function helperGetCaretPosition(wdioElement) { //FIXME Find a way to allow using helper functions inside webdriver.io `execute()` blocks
+ console.log('wdioElement:', wdioElement); //DEBUG
+ // console.log('this:', this); //DEBUG
+ const selector = wdioElement.selector;
+ console.log('selector:', selector); //DEBUG
 
-    const element = document.querySelector(selector);
-    console.log('element.selectionStart:', element.selectionStart); //DEBUG 
-    return element.selectionStart;
-}
-*/
+ const element = document.querySelector(selector);
+ console.log('element.selectionStart:', element.selectionStart); //DEBUG
+ return element.selectionStart;
+ }
+ */
 
 
 //-----------------------------------------------------------------------------
@@ -118,10 +121,10 @@ describe('webdriver.io runner', () => {
         expect(browser.getValue(selectors.inputClassic)).toEqual('987654321te01ststYES!ring');
 
         /*
-        expect(helperGetCaretPosition(inputClassic)).toEqual(42); //FIXME This cannot be called correctly
-        const result = browser.getCaretPosition(inputClassic); //FIXME This cannot be called correctly
-        expect(result).toEqual(19);
-        */
+         expect(helperGetCaretPosition(inputClassic)).toEqual(42); //FIXME This cannot be called correctly
+         const result = browser.getCaretPosition(inputClassic); //FIXME This cannot be called correctly
+         expect(result).toEqual(19);
+         */
 
         // Hold some modifier keys
         browser.keys('End');
@@ -730,5 +733,98 @@ xdescribe('Issue #303', () => { //FIXME Finish this
             return input.selectionStart;
         }).value;
         expect(inputCaretPosition).toEqual(0);
+    });
+});
+
+describe('Issue #403', () => {
+    it('should tests for default values', () => {
+        browser.url(testUrl);
+
+        expect(browser.getValue(selectors.issue403a)).toEqual('25.00%');
+    });
+
+    it('should change the input value accordingly when focusing on the element', () => { //FIXME à tester
+        const inputA = $(selectors.issue403a);
+        const inputB = $(selectors.issue403b);
+
+        // Focus in the input
+        inputA.click();
+
+        // Test the input value while the element is focused
+        expect(browser.getValue(selectors.issue403a)).toEqual('0.25');
+
+        // Focus out of the input
+        inputB.click();
+
+        // Test the input value while the element is not focused
+        expect(browser.getValue(selectors.issue403a)).toEqual('25.00%');
+
+        // Then we cycle back twice just to make sure the value stays the same while tabbing in/out
+        inputA.click();
+        expect(browser.getValue(selectors.issue403a)).toEqual('0.25');
+        inputB.click();
+        expect(browser.getValue(selectors.issue403a)).toEqual('25.00%');
+        inputA.click();
+        expect(browser.getValue(selectors.issue403a)).toEqual('0.25');
+        inputB.click();
+        expect(browser.getValue(selectors.issue403a)).toEqual('25.00%');
+    });
+
+    it('should change the input value accordingly when focusing on the element', () => { //FIXME à tester
+        const inputA = $(selectors.issue403a);
+        const inputB = $(selectors.issue403b);
+
+        // Focus in the input
+        inputB.click();
+        //browser.keys(['Home', 'Delete', 'Delete', 'Delete', 'Delete', '0.12']);
+
+        // Test the input value while the element is focused
+        //expect(browser.getValue(selectors.issue403b)).toEqual('0.012');
+
+        // Focus out of the input
+        inputA.click();
+
+        // Test the input value while the element is not focused
+        expect(browser.getValue(selectors.issue403b)).toEqual('1.20%');
+
+        // Then we cycle back twice just to make sure the value stays the same while tabbing in/out
+        inputB.click();
+        expect(browser.getValue(selectors.issue403b)).toEqual('0.012');
+        inputA.click();
+        expect(browser.getValue(selectors.issue403b)).toEqual('1.20%');
+
+        inputB.click();
+        expect(browser.getValue(selectors.issue403b)).toEqual('0.012');
+        inputA.click();
+        expect(browser.getValue(selectors.issue403b)).toEqual('1.20%');
+    });
+
+    it('should change the input value accordingly when focusing on the element, with a bigger number of decimal places', () => { //FIXME à tester
+        const inputB = $(selectors.issue403b);
+        const inputC = $(selectors.issue403c);
+
+        // Focus in the input
+        inputC.click();
+        browser.keys(['1234567.89']);
+
+        // Test the input value while the element is focused
+        expect(browser.getValue(selectors.issue403c)).toEqual('1,234,567.89');
+
+        // Focus out of the input
+        inputB.click();
+
+        // Test the input value while the element is not focused
+        expect(browser.getValue(selectors.issue403c)).toEqual('1.23457MM');
+
+        // Then we cycle back twice just to make sure the value stays the same while tabbing in/out
+        inputC.click();
+        expect(browser.getValue(selectors.issue403c)).toEqual('1,234,567.89');
+        inputB.click();
+        expect(browser.getValue(selectors.issue403c)).toEqual('1.23457MM');
+
+        inputC.click();
+        expect(browser.getValue(selectors.issue403c)).toEqual('1,234,567.89');
+        inputB.click();
+        expect(browser.getValue(selectors.issue403c)).toEqual('1.23457MM');
     });
 });